requirements.txt
asgiref==3.7.2
Django==5.0
django-admin-thumbnails==0.2.8
django-ckeditor==6.7.0
django-js-asset==2.2.0
django-mptt==0.15.0
gunicorn==21.2.0
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
python-slugify==8.0.1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
